requirements.txt
pyannote.audio
transformers
torch
pydub
datasets
ffmpeg